requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.4
Pillow==6.2.1
pygame==1.9.6
pyparsing==2.4.5
python-dateutil==2.8.1
six==1.13.0
